requirements.txt
streamlit
pdfplumber
pillow
pytesseract
python-dateutil
regex
pytest
